Added phone validator to check phone number and phone input mask moved to models
