+++ composeApp/src/commonMain/kotlin/band/effective/office/elevator/ui/models/PhoneValidator.kt
@@ -1,0 +1,7 @@
+package band.effective.office.elevator.ui.models
+
+abstract class PhoneValidator() {
+    companion object {
+        fun validatePhoneNumber(phoneNumber: String) = phoneNumber.length == 10
+    }
+}
